### 2020-07-09 #2460 @whedon
[PRE REVIEW]: tidyHeatmap: an R package for modular heatmap production based on tidy principles
**Submitting author:** @stemangiola (<a href="http://orcid.org/0000-0001-7474-836X">Stefano Mangiola</a>)
**Repository:** <a href="https://github.com/stemangiola/tidyHeatmap" target ="_blank">https://github.com/stemangiola/tidyHeatmap</a>
**Version:** v1.1.0
