### 2020-07-09 #2460 @kyleniemeyer
Hi @stemangiola,  thanks for your submission to JOSS. As we are getting started, could you please resolve a few issues?
- Please revise your paper to include an explicit Statement of Need section, as described in [our documentation](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)?
- We cannot find an open-source license in your repository.
- Also, it looks like some DOIs may be missing for some references.

Thanks!
