+++ paper/paper.bib
@@ -57,7 +57,8 @@
   pages    = "2847--2849",
   month    =  sep,
   year     =  2016,
-  language = "en"
+  language = "en",
+  doi = {10.1093/bioinformatics/btw313}
 }
 
 @ARTICLE{Kolde2012-tu,
@@ -95,4 +96,5 @@
   issue = {19},
   pages = {2811-2812},
   year = {2014},
+  doi = {10.1093/bioinformatics/btu393}
 }
