### 2020-07-10 268b106
small fixes to repo
