+++ DESCRIPTION
@@ -17,7 +17,7 @@ Description: This is a tidy implementation for heatmap.  At the
     a grouped tbl. For example: df %>% group_by(...).  Labels size
     adjusted by row and column total number.  Default use of Brewer and
     Viridis palettes.
-License: GPL-3
+License: GPL-3 + file LICENSE
 URL: https://www.r-project.org,
     https://github.com/stemangiola/tidyHeatmap
 BugReports: https://github.com/stemangiola/tidyHeatmap
