### 2020-07-10 #2460 @stemangiola
The statement of need has been added to paper.md
