### 2020-07-11 #2472 @whedon
[REVIEW]: tidyHeatmap: an R package for modular heatmap production based on tidy principles
**Submitting author:** @stemangiola (<a href="http://orcid.org/0000-0001-7474-836X">Stefano Mangiola</a>)
**Repository:** <a href="https://github.com/stemangiola/tidyHeatmap" target ="_blank">https://github.com/stemangiola/tidyHeatmap</a>
**Version:** v1.1.2
