+++ man/heatmap-methods.Rd
@@ -5,6 +5,8 @@
 \alias{heatmap}
 \alias{heatmap_}
 \alias{heatmap,tbl-method}
+\alias{heatmap,tbl_df-method}
+\alias{heatmap,tidybulk-method}
 \title{Creates a  `InputHeatmap` object from `tbl_df` on evaluation creates a `ComplexHeatmap`}
 \usage{
 heatmap(
@@ -69,6 +71,48 @@ heatmap_(
   palette_abundance = NULL,
   ...
 )
+
+\S4method{heatmap}{tbl_df}(
+  .data,
+  .row,
+  .column,
+  .value,
+  transform = NULL,
+  .scale = "row",
+  palette_value = c("#440154FF", "#21908CFF", "#fefada"),
+  palette_grouping = list(),
+  annotation = NULL,
+  type = rep("tile", length(quo_names(annotation))),
+  palette_discrete = list(),
+  palette_continuous = list(),
+  .abundance = NULL,
+  .horizontal = NULL,
+  .vertical = NULL,
+  log_transform = NULL,
+  palette_abundance = NULL,
+  ...
+)
+
+\S4method{heatmap}{tidybulk}(
+  .data,
+  .row,
+  .column,
+  .value,
+  transform = NULL,
+  .scale = "row",
+  palette_value = c("#440154FF", "#21908CFF", "#fefada"),
+  palette_grouping = list(),
+  annotation = NULL,
+  type = rep("tile", length(quo_names(annotation))),
+  palette_discrete = list(),
+  palette_continuous = list(),
+  .abundance = NULL,
+  .horizontal = NULL,
+  .vertical = NULL,
+  log_transform = NULL,
+  palette_abundance = NULL,
+  ...
+)
 }
 \arguments{
 \item{.data}{A `tbl_df` formatted as | <ELEMENT> | <FEATURE> | <VALUE> | <...> |}
@@ -109,6 +153,10 @@ heatmap_(
 }
 \value{
 A `InputHeatmap` objects that gets evaluated to a `ComplexHeatmap` object
+
+A `InputHeatmap` object
+
+A `InputHeatmap` object
 
 A `InputHeatmap` object
 
