### 2020-07-13 ec2417f
fixed tile palette bug if df grouped
