+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: tidyHeatmap
 Title: A Tidy Implementation of Heatmap
-Version: 1.1.0
+Version: 1.1.1
 Authors@R: 
     person(given = "Stefano",
            family = "Mangiola",
