### 2020-07-14 stemangiola/tidyHeatmap#38 @rlbarter
Issue installing package
Hello!

I'm currently reviewing your package submission to JOSS and am having difficulty installing your package. Not sure what the issue is, but I won't be able to proceed with the review until I can install the package :(.

Here is the output when I try to install:

```
> devtools::install_github("stemangiola/tidyHeatmap")
Using github PAT from envvar GITHUB_PAT
Downloading GitHub repo stemangiola/tidyHeatmap@master
✓  checking for file ‘/private/var/folders/yv/wgyt24cx34j3fyxz_y4lzx380000gn/T/RtmpBkVdRc/remotes2d4d352e98ee/stemangiola-tidyHeatmap-0f9fafa/DESCRIPTION’ ...
─  preparing ‘tidyHeatmap’:
✓  checking DESCRIPTION meta-information ...
─  installing the package to process help pages
         -----------------------------------
─  installing *source* package ‘tidyHeatmap’ ...
   ** using staged installation
   ** R
   ** data
   *** moving datasets to lazyload DB
   ** inst
   ** byte-compile and prepare package for lazy loading
   Error in dyn.load(file, DLLpath = DLLpath, ...) :
     unable to load shared object '/Library/Frameworks/R.framework/Versions/4.0/Resources/library/Rcpp/libs/Rcpp.so':
     dlopen(/Library/Frameworks/R.framework/Versions/4.0/Resources/library/Rcpp/libs/Rcpp.so, 6): Symbol not found: _EXTPTR_PTR
     Referenced from: /Library/Frameworks/R.framework/Versions/4.0/Resources/library/Rcpp/libs/Rcpp.so
     Expected in: /Library/Frameworks/R.framework/Versions/4.0/Resources/lib/libR.dylib
    in /Library/Frameworks/R.framework/Versions/4.0/Resources/library/Rcpp/libs/Rcpp.so
   Calls: <Anonymous> ... asNamespace -> loadNamespace -> library.dynam -> dyn.load
   Execution halted
   ERROR: lazy loading failed for package ‘tidyHeatmap’
─  removing ‘/private/var/folders/yv/wgyt24cx34j3fyxz_y4lzx380000gn/T/RtmpTWD0jE/Rinstbdc9236f7c8/tidyHeatmap’
         -----------------------------------
   ERROR: package installation failed
Error: Failed to install 'tidyHeatmap' from GitHub:
  System command 'R' failed, exit status: 1, stdout + stderr (last 10 lines):
E>   dlopen(/Library/Frameworks/R.framework/Versions/4.0/Resources/library/Rcpp/libs/Rcpp.so, 6): Symbol not found: _EXTPTR_PTR
E>   Referenced from: /Library/Frameworks/R.framework/Versions/4.0/Resources/library/Rcpp/libs/Rcpp.so
E>   Expected in: /Library/Frameworks/R.framework/Versions/4.0/Resources/lib/libR.dylib
E>  in /Library/Frameworks/R.framework/Versions/4.0/Resources/library/Rcpp/libs/Rcpp.so
E> Calls: <Anonymous> ... asNamespace -> loadNamespace -> library.dynam -> dyn.load
E> Execution halted
E> ERROR: lazy loading failed for package ‘tidyHeatmap’
E> * removing ‘/private/var/folders/yv/wgyt24cx34j3fyxz_y4lzx380000gn/T/RtmpTWD0jE/Rinstbdc9236f7c8/tidyHeatmap’
E>       -----------------------------------
E> ERROR: package installation failed
```

Here is my session information:

```
> sessionInfo()
R version 4.0.0 (2020-04-24)
Platform: x86_64-apple-darwin17.0 (64-bit)
Running under: macOS High Sierra 10.13.6

Matrix products: default
BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib
LAPACK: /Library/Frameworks/R.framework/Versions/4.0/Resources/lib/libRlapack.dylib

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

attached base packages:

other attached packages:
 [1] patchwork_1.0.0  conflicted_1.0.4 lubridate_1.7.9  scales_1.1.1     yardstick_0.0.6  forcats_0.5.0
 [7] stringr_1.4.0    dplyr_1.0.0.9000 purrr_0.3.4      readr_1.3.1      tidyr_1.1.0      tibble_3.0.3
[13] ggplot2_3.3.2    tidyverse_1.3.0  gt_0.2.1         L1pack_0.38.19   MASS_7.3-51.5    kableExtra_1.1.0
[19] knitr_1.28

 [1] nlme_3.1-147      fs_1.4.1          usethis_1.6.1     devtools_2.3.0    webshot_0.5.2     httr_1.4.1
 [7] rprojroot_1.3-2   tools_4.0.0       backports_1.1.8   R6_2.4.1          DBI_1.1.0         mgcv_1.8-31
[13] colorspace_1.4-1  withr_2.2.0       tidyselect_1.1.0  gridExtra_2.3     prettyunits_1.1.1 processx_3.4.3
[19] curl_4.3          compiler_4.0.0    cli_2.0.2         rvest_0.3.5       xml2_1.3.2        desc_1.2.0
[25] labeling_0.3      bookdown_0.19.1   callr_3.4.3       digest_0.6.25     rmarkdown_2.2.5   pkgconfig_2.0.3
[31] htmltools_0.5.0   sessioninfo_1.1.1 dbplyr_1.4.3      rlang_0.4.7       readxl_1.3.1      rstudioapi_0.11
[37] farver_2.0.3      generics_0.0.2    jsonlite_1.6.1    magrittr_1.5      Matrix_1.2-18     Rcpp_1.0.5
[43] munsell_0.5.0     fansi_0.4.1       lifecycle_0.2.0   stringi_1.4.6     pROC_1.16.2       yaml_2.2.1
[49] pkgbuild_1.1.0    plyr_1.8.6        grid_4.0.0        crayon_1.3.4      lattice_0.20-41   haven_2.2.0
[55] splines_4.0.0     hms_0.5.3         ps_1.3.3          pillar_1.4.6      pkgload_1.1.0     reprex_0.3.0
[61] glue_1.4.1        evaluate_0.14     remotes_2.1.1     modelr_0.1.6      vctrs_0.3.1       testthat_2.3.2
[67] cellranger_1.1.0  gtable_0.3.0      assertthat_0.2.1  xfun_0.14         broom_0.5.6       viridisLite_0.3.0
[73] tinytex_0.23      memoise_1.1.0     ellipsis_0.3.1
```

Best,
Rebecca
