### 2020-07-14 #2472 @rlbarter
Created an issue in target repository due to installation issues: https://github.com/stemangiola/tidyHeatmap/issues/38
