### 2020-07-14 stemangiola/tidyHeatmap#38 @rlbarter
After further exploration, I think this issue was on my end. My laptop crashed earlier today and weird things were happening all over. Updated R version, and can install fine now!
