### 2020-07-14 #2472 @rlbarter
Issue resolved - was my laptop acting funky due to crashing earlier. A restart solved the problem.
