### 2020-07-14 #2472 @rlbarter
I've completed my review.

This looks great overall.

I have a few very minor comments for the authors, which they can feel free to take or disregard.

- I noticed a minor typo in contribution section of README: "**pleae** open an issue" --> "**please** open an issue". Maybe just do a full spell-check?

- In the README, consider updating `gather` to `pivot_longer` just to be up-to-date with tidyverse function.

- In the README, use `->` instead of `=`, just to follow usual convention.

- My heatmap package superheat was one of the early packages to add point, bar, line annotation to heatmaps, so it would be a relevant citation (https://www.tandfonline.com/doi/abs/10.1080/10618600.2018.1473780). Your package works much better with the modern tidyverse style (mine was built way back in 2015) - nicely done.
