### 2020-07-14 #2472 @stemangiola
Thanks a lot @rlbarter all four suggestions have been incorporated in this [pull request](https://github.com/stemangiola/tidyHeatmap/pull/39). Your package have great graphics, probably the cleanest of all.
