### 2020-07-16 6f8c057
make sd even more robust
