+++ man/type_to_annot_function.Rd
@@ -1,16 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/data.R
-\docType{data}
-\name{type_to_annot_function}
-\alias{type_to_annot_function}
-\title{type_to_annot_function annot_*}
-\format{
-An object of class \code{list} of length 4.
-}
-\usage{
-type_to_annot_function
-}
-\description{
-type_to_annot_function annot_*
-}
-\keyword{datasets}
