### 2020-07-19 #2472 @AliciaSchep
Great work -- really nice interface for making heatmaps from dataframes!

In terms of the criteria for review I saw two main areas for potential improvement --

(1) The tests appear to mostly check that the functions run without error and create an output of the right class.  This strategy seems like it could miss many potential issues. One package that can be helpful for visual regression testing is [vdiffr](https://github.com/r-lib/vdiffr).  Or perhaps since this package is relying on ComplexHeatmap, testing for the actual rendered image is a step too far, but there could be tests to make sure that `InputHeatmap` objects get correctly converted into the right objects from ComplexHeatmap.

(2) I think both the paper and the documentation could benefit from a bit more explanation of how the various parts of a complex heatmap map to the "tidy" data framework.  For example in the `mtcars` example, why do the `hp` and `vs` variables not get pulled into the `Property` column? My first instinct for tidying that data frame would be that all variables get pulled into `Property`. It seems like the difference with those in terms of how you structure the data is in how you will want to use the variable in the heatmap. Maybe thats kind of an arbitrary example trying to make use of a common data set, but I think given that a major contribution of this work is providing a bridge between "tidy" data and heatmaps it would be helpful to have a bit more explanation of how heatmaps and this API relate to the data and its structure.  In the paper it states that "Tidy data frames are characterized by having a specific structure where each variable is a column and each observation is a row.".  It'd be helpful to expand on how the tidyHeatmap expects "annotation" data to be included in such data and how that fits the 'tidy' paradigm.  It might also be useful to compare the tidy representation with representations used in Bioconductor like SummarizedExperiment which in some ways more obviously map to the heatmap visualization.
