+++ tests/testthat/tests.R
@@ -1,5 +1,7 @@
 context('tests')
 
+# For resetting
+# vdiffr::manage_cases()
 
 test_that("basic plot",{
 
@@ -11,8 +13,8 @@ test_that("basic plot",{
 			.value = `read count normalised log`
 		)
 	
+	vdiffr::expect_doppelganger("base", p)
 	
-  expect_equal(as.character(class(p)), "InputHeatmap" )
 
 })
 
@@ -30,7 +32,7 @@ test_that("grouped plot",{
 		)
 	
 	
-	expect_equal(as.character(class(p)), "InputHeatmap" )
+	vdiffr::expect_doppelganger("grouped", p)
 	
 })
 
@@ -45,14 +47,13 @@ test_that("annotated plot numerical continuous intereg nominal annot",{
 		) %>%
 		add_tile(CAPRA_TOTAL)
 	
-	expect_equal(as.character(class(p)), "InputHeatmap" )
+	vdiffr::expect_doppelganger("annotated heatmap 1", p)
 	
 })
 
 test_that("annotated plot continuous annot MUST ERROR",{
 	
 	my_df = dplyr::filter(tidyHeatmap::N52, Category == "Angiogenesis")
-	
 	expect_error(
 		tidyHeatmap::heatmap(
 			 left_join(my_df,  dplyr::mutate(dplyr::distinct(my_df, sample), a = rnorm(n()))), 
@@ -67,10 +68,12 @@ test_that("annotated plot continuous annot MUST ERROR",{
 test_that("annotated plot continuous annot as well",{
 	
 	my_df = dplyr::filter(tidyHeatmap::N52, Category == "Angiogenesis")
+	my_UBR = c( -0.4271163,  5.3530527, -0.7269678, -5.8277242, -4.0925786,  3.4246804, -1.6002821, -6.5576121,  -2.9980416 ,-0.6470534,  4.4336807, -0.7569798,  1.6489560)
 	
+	set.seed(123)
 	p = 
 		tidyHeatmap::heatmap(
-			left_join(my_df,  dplyr::mutate(dplyr::distinct(my_df, UBR), a = rnorm(n(), sd=5))), 
+			left_join(my_df,  dplyr::mutate(dplyr::distinct(my_df, UBR), a = my_UBR)), 
 			.column = UBR, 
 			.row = symbol_ct, 
 			.value = `read count normalised log`
@@ -78,7 +81,7 @@ test_that("annotated plot continuous annot as well",{
 		add_tile(a) %>%
 		add_tile(CAPRA_TOTAL)
 	
-	expect_equal(as.character(class(p)), "InputHeatmap" )
+	vdiffr::expect_doppelganger("annotated heatmap 2", p)
 	
 })
 
@@ -97,7 +100,7 @@ test_that("grouped and annotated plot",{
 		add_tile(CAPRA_TOTAL)
 	
 	
-	expect_equal(as.character(class(p)), "InputHeatmap" )
+	vdiffr::expect_doppelganger("grouped annotated heatmap 1", p)
 	
 })
 
@@ -114,7 +117,7 @@ test_that("grouped double and annotated plot",{
 		add_tile(activation)
 	
 	
-	expect_equal(as.character(class(p)), "InputHeatmap" )
+	vdiffr::expect_doppelganger("grouped annotated heatmap 2", p)
 	
 	
 })
@@ -150,7 +153,7 @@ test_that("pasilla one annotation",{
 		add_tile(condition)
 	
 	
-	expect_equal(as.character(class(p)), "InputHeatmap" )
+	vdiffr::expect_doppelganger("pasilla heatmap 1", p)
 	
 })
 
[…]
