+++ DESCRIPTION
@@ -46,6 +46,7 @@ Imports:
     lifecycle (>= 0.2.0)
 Suggests:
     testthat,
+    vdiffr,
     BiocManager,
     knitr,
     rmarkdown,
