### 2020-07-20 b0192de
improve documentation and article for data structure
