### 2020-07-20 9a4bb87
solve discrepancy Max Win Linux for log1p
