### 2020-07-20 #2472 @stemangiola
Thanks Alicia.

Both your issue and feedback have been integrated here https://github.com/stemangiola/tidyHeatmap/pull/44

1) Great suggestion for vdiffr! I wondered how to test images. That has been integrated
2) Good suggestion. I realised that while is genetics the structure element/feature/value/independent variable is almost implicit, this can create confusion generally. I have update both README, vignette and the article, better explaining (also with examples) what there four+ columns mean.
