### 2020-07-20 #2472 @stemangiola
And for commenting on your concern

"Maybe thats kind of an arbitrary example trying to make use of a common data set, but I think given that a major contribution of this work is providing a bridge between "tidy" data and heatmaps it would be helpful to have a bit more explanation of how heatmaps and this API relate to the data and its structure. "

Yes this was totally arbitrary in order to produce an heatmap with annotations. It can be confusing, and I tried to clarify substantially in the README and vignette.
