### 2020-07-20 6c061e8
void fails for max and win, unknown reason. plot don't match
