### 2020-07-20 ec4fd34
try to solve tbl inheritance (e.g. tidybulk)
