### 2020-07-25 d6ab1e5
setup spelling check
