+++ DESCRIPTION
@@ -45,6 +45,7 @@ Imports:
     grDevices,
     lifecycle (>= 0.2.0)
 Suggests:
+    spelling,
     testthat,
     vdiffr,
     BiocManager,
@@ -62,3 +63,4 @@ biocViews: AssayDomain, Infrastructure
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.0
+Language: en-US
