### 2020-07-25 #2472 @stemangiola
>
>
> Incorporating `vdiffr` addresses my concerns about the tests and the proposed updates to readme look good -- although I would recommend running a spellchecker as I noticed a couple spelling issues. There's an R package that can help with that -- [spelling](https://github.com/ropensci/spelling)

Thanks Alicia, I have updated the pull request with the spell checking framework, and corrected spelling.
