### 2020-07-27 #2472 @karthik
Thanks @stemangiola

Let me know once you've merged the spell checks. I found some more (see below) but you may have addressed those in the issues. Be sure to address:

```
DESCRIPTION does not contain 'Language' field. Defaulting to 'en-US'.
```

 <summary>Possible spelling errors in the package</summary>
<pre><code>
DESCRIPTION does not contain 'Language' field. Defaulting to 'en-US'.
  WORD              FOUND IN
acceots           ifelse_pipe.Rd:5,22
                  ifelse2_pipe.Rd:5,26
annot             annot_to_list.Rd:5,16
centred           scale_robust.Rd:13
circlize          heatmap-methods.Rd:130
colorRamp         heatmap-methods.Rd:130
                  input_heatmap.Rd:32
                  README.md:187
                  README.Rmd:130
colour            description:4
colous            input_heatmap.Rd:32
ComplexHeatmap    add_annotation.Rd:27,30
                  add_bar.Rd:5,17,25
                  add_line.Rd:5,17,25
                  add_point-methods.Rd:7,21,31
                  add_tile-methods.Rd:7,21,31
                  heatmap-methods.Rd:10,152,155,166,171
                  input_heatmap.Rd:36,39,42
                  save_pdf-methods.Rd:31
                  description:2
                  README.md:21
                  README.Rmd:18
                  introduction.Rmd:26
customisation     input_heatmap.Rd:32
df                add_bar-methods.Rd:11
                  add_bar.Rd:10
                  add_line-methods.Rd:11
                  add_line.Rd:10
                  add_point-methods.Rd:14
                  add_tile-methods.Rd:14
                  heatmap-methods.Rd:10,118
                  description:7
dplyr             add_annotation.Rd:30
                  heatmap-methods.Rd:166
                  input_heatmap.Rd:42
                  README.md:157
                  README.Rmd:107
enquo             get_abundance_norm_if_exists.Rd:15
                  get_elements_features_abundance.Rd:27
                  get_elements_features.Rd:19
                  get_elements.Rd:17
                  get_sample_counts.Rd:17
                  get_sample_transcript_counts.Rd:19
                  get_sample_transcript.Rd:17
generalisation    ifelse_pipe.Rd:5,22
                  ifelse2_pipe.Rd:5,26
github            README.md:17
                  README.Rmd:15
ifelse            ifelse_pipe.Rd:5,22
                  ifelse2_pipe.Rd:5,26
InputHeatmap      add_annotation.Rd:16
                  add_bar.Rd:5,17,20
                  add_line.Rd:5,17,20
                  add_point-methods.Rd:7,21,26
                  add_tile-methods.Rd:7,21,26
                  heatmap-methods.Rd:10,155,157,159,161,163,171
io                README.md:17
                  README.Rmd:15
lifecycle         README.md:51
                  README.Rmd:49
                  introduction.Rmd:56
Lifecycle         README.md:6
                  README.Rmd:7
mtcars            README.md:136
                  README.Rmd:89
Pasilla           pasilla.Rd:6,14
pleae             README.md:60
                  README.Rmd:59
quosure           quo_names.Rd:5,16
quosures          quo_names.Rd:10
rlang             prepend.Rd:5,20
sd                scale_robust.Rd:5,16
stemangiola       README.md:17
                  README.Rmd:15
tbl               add_annotation.Rd:30
                  add_bar-methods.Rd:11
                  add_bar.Rd:10
                  add_line-methods.Rd:11
                  add_line.Rd:10
                  add_point-methods.Rd:14
                  add_tile-methods.Rd:14
                  check_if_counts_is_na.Rd:15
                  check_if_duplicated_genes.Rd:19
                  check_if_wrong_input.Rd:17
                  get_x_y_annotation_columns.Rd:10
                  heatmap-methods.Rd:10,118,166,171
                  input_heatmap.Rd:20,42
                  description:7
                  README.md:27
                  README.Rmd:23
                  introduction.Rmd:31
tibble            add_annotation.Rd:22,24,30
                  add_attr.Rd:10,17
                  add_bar-methods.Rd:15
                  add_bar.Rd:14
                  add_class.Rd:10,15
                  add_line-methods.Rd:15
                  add_line.Rd:14
                  add_point-methods.Rd:18
                  add_tile-methods.Rd:18
                  as_matrix.Rd:5,10,20
                  check_if_counts_is_na.Rd:10
                  check_if_duplicated_genes.Rd:10
                  check_if_wrong_input.Rd:10
                  drop_class.Rd:10,15
                  get_abundance_norm_if_exists.Rd:10
                  get_elements_features_abundance.Rd:16
                  get_elements_features.Rd:10
                  get_elements.Rd:10
                  get_sample_counts.Rd:10
                  get_sample_transcript_counts.Rd:10
                  get_sample_transcript.Rd:10
                  heatmap-methods.Rd:166
                  ifelse_pipe.Rd:10,19
                  ifelse2_pipe.Rd:10,23
                  input_heatmap.Rd:42
                  scale_design.Rd:10,15
                  select_closest_pairs.Rd:10,13
                  README.md:9
                  README.Rmd:12
tidybulk          README.md:14
                  README.Rmd:13
tidygate          README.md:12
                  README.Rmd:12
tranform          input_heatmap.Rd:28
transcriptomics   README.md:9
                  README.Rmd:13
viridis           heatmap-methods.Rd:130
Viridis           description:9
                  README.md:30
                  README.Rmd:26
                  introduction.Rmd:33
visualisation     add_bar.Rd:25
                  add_line.Rd:25
                  add_point-methods.Rd:31
                  add_tile-methods.Rd:31
</code>
</pre>
