+++ paper/paper.md
@@ -43,16 +43,15 @@ Recently, efforts have been made toward the harmonisation of data frame structur
 Considering (i) the utility and complexity of creating information-rich heatmaps, and (ii) the opportunity of increased coding efficiency and robustness offered by the tidy paradigm, a bridge between the two is very much needed. Recently, many tools for data science have been implemented according with tidy principles, this package aims to fill the gap for one of the most used data explorations tool.
 
 # Tidy paradigm for visualisation
-`tidyHeatmap` is a graphical R package that introduces tidy principles to the creation of information-rich heatmaps. It is available in the CRAN R repository. This package currently uses `ComplexHeatmap` as its graphical engine; however, due to its modular design it can be readily expanded to interface other engines. The command-line user interface is organised into (i) a main plotting utility; (ii) annotation layer utilities; and (iii) file output utilities. The input is a tidy data frame with element (e.g. person), feature (e.g. physical characteristics) and value column, with additional column for independent variables for either elements (e.g. number of sport medals) or features (e.g. macroscopic or molecular characteristics). In this data structure, each observation is an element-feature pair. The input data frame streams along the utility path using the pipe operator from `magrittr`, allowing a high-level of modularity. The main utility allows the user to plot a base heatmap with dendrograms. The annotation utilities allow the user to serially add tile, point, bar and/or line annotation boxes to the side of the heatmap. The orientation of the annotations (row- or column-wise) is inferred by the `tidyHeatmap` algorithms, based on the input data frame. The file output utility allows the user to write vector or bitmap images directly from the R object, in the style of `ggplot2`. User defined row- or column-wise clusters can be defined effortlessly by applying the `group_by` function from `dplyr` [@Hadley:2020] to the input data frame. Data transformation, row and column scaling is done internally. Together, this leads to a decrease of coding burden of 3 and 5 folds for lines and characters respectively compared to `ComplexHeatmap` (e.g., for \autoref{fig:example}). Besides offering a modular and user-friendly interface, `tidyHeatmap` applies publication-ready aesthetics such as `viridis` [@Garnier:2018] and `brewer` [@Neuwirth:2014] colour palettes and automatic sizing of row and column labels to avoid overlapping (\autoref{fig:example}). 
-
-![Heatmap of the pasilla dataset including grouping and multiple annotations. Some annotation data was simulated for visualisation purposes. \label{fig:example}](paper_tables_and_figures_files/figure-gfm/example_figure-1.png)
-
-The input to `tidyHeatmap` is a tidy data frame with three basic columns including row and column element identifiers, and values that will be converted into colours. Additionally, further columns can include information about element grouping and annotation.
+`tidyHeatmap` is a graphical R package that introduces tidy principles to the creation of information-rich heatmaps. It is available in the CRAN R repository. This package currently uses `ComplexHeatmap` as its graphical engine; however, due to its modular design it can be readily expanded to interface other engines. The command-line user interface is organised into (i) a main plotting utility; (ii) annotation layer utilities; and (iii) file-output utilities. The input is a tidy data frame with element (e.g. person), feature (e.g. physical characteristics) and value column, with additional column for independent variables for either elements (e.g. number of sport medals) or features (e.g. macroscopic or molecular characteristics). In this data structure, each observation is an element-feature pair. 
 
 | element         | feature         | value     | annotation | group |
 | --------------- | --------------- | --------- | ---------- | ----- |
 | `chr` or `fctr` | `chr` or `fctr` | `numeric` | …          | …     |
 
[…]
