### 2020-07-27 06aa2b9
Update utilities.R
