+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: tidyHeatmap
 Title: A Tidy Implementation of Heatmap
-Version: 1.1.1
+Version: 1.1.2
 Authors@R: 
     c(person(given = "Stefano",
            family = "Mangiola",
