+++ README.md
@@ -2,8 +2,8 @@ tidyHeatmap
 ================
 
 <!-- badges: start -->
-
 [![Lifecycle:maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3966515.svg)](https://doi.org/10.5281/zenodo.3966515)
 <!-- badges: end -->
 
 Please have a look also to -
