### 2020-07-30 #2472 @karthik
@stemangiola Thanks. Can you confirm that the version is now `v1.1.2`?
