### 2020-07-31 #2472 @stemangiola
Hello @karthik all good except few packages names seem to have been printed in two lines. magrittr and ComplexHeatmap are splt across lines with no hyphenation.
