### 2020-07-31 #2472 @kthyng
Hi @stemangiola! I will help finishing publishing your paper. I suggest that you change the metadata in your Zenodo archive to exactly match your JOSS paper title and authors.
