### 2020-08-01 2e86e9a
Tweaks to paper
For JOSS submission [#2472](https://github.com/openjournals/joss-reviews/issues/2472).
