### 2020-08-01 5ef38f0
Fixes reference capitalization
For JOSS submission [#2472](https://github.com/openjournals/joss-reviews/issues/2472).
