+++ paper/paper.bib
@@ -9,7 +9,7 @@
 }
 @Book{Hadley:2016,
   author = {Hadley Wickham},
-  title = {ggplot2: Elegant Graphics for Data Analysis},
+  title = {{ggplot2}: Elegant Graphics for Data Analysis},
   publisher = {Springer-Verlag New York},
   year = {2016},
   isbn = {978-3-319-24277-4},
@@ -68,14 +68,14 @@
   year    =  2012
 }
 @Manual{Hadley:2020,
-  title = {dplyr: A Grammar of Data Manipulation},
+  title = {{dplyr}: A Grammar of Data Manipulation},
   author = {Hadley Wickham and Romain François and Lionel Henry and Kirill Müller},
   year = {2020},
   note = {R package version 1.0.0},
   url = {https://CRAN.R-project.org/package=dplyr},
 }
 @Manual{Garnier:2018,
-  title = {viridis: Default Color Maps from 'matplotlib'},
+  title = {{viridis}: Default Color Maps from 'matplotlib'},
   author = {Simon Garnier},
   year = {2018},
   note = {R package version 0.5.1},
@@ -89,7 +89,7 @@
   url = {https://CRAN.R-project.org/package=RColorBrewer},
 }
 @Article{Zuguang:2014,
-  title = {circlize implements and enhances circular visualization in R},
+  title = {circlize implements and enhances circular visualization in {R}},
   author = {Zuguang Gu and Lei Gu and Roland Eils and Matthias Schlesner and Benedikt Brors},
   journal = {Bioinformatics},
   volume = {30},
@@ -100,7 +100,7 @@
 }
 @article{superheatmap,
 author = {Rebecca L. Barter and Bin Yu},
-title = {Superheat: An R Package for Creating Beautiful and Extendable Heatmaps for Visualizing Complex Data},
+title = {Superheat: An {R} Package for Creating Beautiful and Extendable Heatmaps for Visualizing Complex Data},
 journal = {Journal of Computational and Graphical Statistics},
 volume = {27},
 number = {4},
