### 2020-08-03 #2472 @kthyng
the doi is still resolving but in the meantime...

Congrats to  @stemangiola on your new publication! Thanks to editor @karthik and reviewers @rlbarter and @AliciaSchep — this process wouldn't be possible without your expertise and time.
