### 2020-08-03 #2472 @danielskatz
@openjournals/dev  - for me, the DOI resolves but the PDF does not load in the frame where it should - any thoughts?
